refactor: dropbox 컴포넌트 header/profile에서 사용되므로 폴더 이동 로그아웃 버튼 분자로 분리 및 로그아웃 상태 분리
